pyproject.toml
[project]
name = "acit-2911-agile-development-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "blinker>=1.9.0",
    "flask-login>=0.6.3",
    "flask-pymongo>=3.0.1",
    "flask-wtf>=1.3.0",
    "pymongo>=4.17.0",
    "werkzeug>=3.1.8",
    "wtforms>=3.2.1",
]
